requirements.txt
contourpy==1.1.1
cycler==0.12.0
fonttools==4.43.0
googleapis-common-protos==1.60.0
grpclib==0.4.5
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
kiwisolver==1.4.5
matplotlib==3.8.0
multidict==6.0.4
numpy==1.26.0
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.1
protobuf==4.24.3
pyparsing==3.1.1
python-dateutil==2.8.2
scipy==1.11.3
six==1.16.0
svgpathtools==1.6.1
svgwrite==1.4.3
tk==0.1.0
typing_extensions==4.8.0
viam-sdk==0.6.1
